requirements.txt
Django~=5.2.1
djangorestframework~=3.16.0
python-dotenv==1.0.0
Pillow==10.3.0